% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affs.R
\name{affs}
\alias{affs}
\title{affs Function}
\usage{
affs(x, na.action = na.omit)
}
\arguments{
\item{x}{a matrix where the rows represent samples, and the columns represents the features}
}
\value{
affinity graph
}
\description{
constructing  local scaling affinity
}
