% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.num.graph.R
\name{es.num.graph}
\alias{es.num.graph}
\title{es.num.graph.}
\usage{
es.num.graph(W, NUMC = 2:8)
}
\arguments{
\item{W}{the affinity graph.}

\item{NUMC}{a vector which contains the possible choices of number of clusters.}
}
\value{
The estimated number of clusters according to rotation cost.

Values of rotation cost.
}
\description{
This function estimates the optimal number of clusters given graph.
}
