% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSD.R
\name{MOSD}
\alias{MOSD}
\title{Multi-omics integration with weighted local scaling affinity and self-diffusion  applied  for cancer subtyping}
\usage{
MOSD(x, k = NULL)
}
\arguments{
\item{x}{A list  of matrices contains multi-omics datasets where row is samples and column is features}

\item{k}{Clustering number;if k==NULL,the clustering number is estimated by seperarion cost method.}
}
\value{
a list;
(1)clu:the clustering result;
(2)S:the final integrated diffusion matrix;
(3)es.num:the estimated clustering number;
}
\description{
Cancer subtypes identifition by multi-omics integration
}
\examples{
data(COADREAD)
d<-list(mydatGE,mydatME,mydatMI)
test <- MOSD(d)
}
