% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.KM.R
\name{plot.KM}
\alias{plot.KM}
\title{plot survival curv}
\usage{
\method{plot}{KM}(
  clinical,
  labels,
  limit = NULL,
  annot = NULL,
  color = NULL,
  xlab = "Follow up",
  ylab = "Survival Probability",
  title = NULL,
  legend.pos = "top",
  palette = "jama_classic",
  risk.table = T,
  risk.table.ratio = 0.4,
  anno.pos = "bottom",
  anno.x.shift = 0.5
)
}
\arguments{
\item{clinical}{Surv(time,event)}

\item{labels}{labels}

\item{color}{color}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\description{
plot survival curv
}
\examples{
not run
clinical<- Surv(as.numeric(time), as.numeric(status))
plot.KM(clinical,labels)
}
