% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self.diffusion.R
\name{self.diffusion}
\alias{self.diffusion}
\title{self.diffusion Function}
\usage{
self.diffusion(A, K)
}
\arguments{
\item{A}{affinity graph}

\item{K}{iteration steps}
}
\value{
diffused graph
}
\description{
Implementing self-diffusion on local scaling affinity
}
\examples{
data(GSE62254_data)
data1<-apply(GSE62254_gene.expression, 2, function(x) (x-mean(x))/sd(x))
A<-affs(data1)
self.diffusion(A,4)
}
\keyword{A}
\keyword{K}
