% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.clu.R
\name{spec.clu}
\alias{spec.clu}
\title{spectralClustering}
\usage{
spec.clu(affinity, K, type = 3)
}
\arguments{
\item{affinity}{input graph}

\item{K}{the number of clusters}

\item{type}{indicators of variants of spectral clustering}
}
\value{
output label
}
\description{
spectral clustering
}
\keyword{K}
\keyword{affinity}
